backend/requirements.txt
# dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
typer==0.9.0
email-validator==2.1.1

# db
sqlalchemy==2.0.23
sqlalchemy_utils==0.41.1
alembic==1.13.0
asyncpg==0.29.0
aiosqlite==0.19.0

# auth & security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP client
httpx==0.25.2

# dev & testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==24.3.0

# terminal process management 
ptyprocess==0.7.0
pexpect==4.8.0
psutil==5.9.8
aiofiles==23.2.1

# websocket support
websockets==12.0

# data science libraries
pandas==2.1.4
scipy==1.11.4
